Put some code in helpers / middleware
